Modularizing project structure for 2.0: reorganizing classes into appropriate packages
