+++ src/main/java/com/victorvalentim/zividomelive/management/LogManager.java
@@ -1,4 +1,4 @@
-package com.victorvalentim.zividomelive;
+package com.victorvalentim.zividomelive.management;
 
 import java.io.IOException;
 import java.util.logging.*;
