+++ src/main/java/com/victorvalentim/zividomelive/rendering/EquirectangularRenderer.java
@@ -1,4 +1,4 @@
-package com.victorvalentim.zividomelive;
+package com.victorvalentim.zividomelive.rendering;
 
 import processing.core.*;
 import processing.opengl.*;
